requirements.txt
# Jal Sarovar - Amrit Sarovar Water Quality Management System
# Requirements

# Flask and Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.10.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security
Werkzeug==3.0.1
Authlib==1.3.0
itsdangerous==2.1.2
Flask-Limiter==3.5.0
redis==5.0.1

# ML and Data Processing
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4
joblib==1.3.2
xgboost==2.0.3

# Utilities
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
